requirements.txt
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.1
dnspython==2.7.0
loguru==0.7.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
pycparser==2.22
pymongo==4.10.1
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dotenv==1.0.1
win32-setctime==1.1.0
